+++ src/assets/help/changelog.md
@@ -1,6 +1,8 @@
 # 2025-08-25
 
 - Updates to ROI and Resource ROI Overview Help texts (by `lowstrife`)
+- Implements the Production Chain Tool [#127](https://github.com/PRUNplanner/frontend/issues/127)
+
 # 2025-08-24
 
 - The backend communication layer has been restructured to be key-based rather than using direct function calls, improving handling and enabling more predictable state management.